+++ packages/sdk/src/swap/types.ts
@@ -78,6 +78,7 @@ interface DepositAddressFields extends SwapStatusResponseCommonFields {
   estimatedDepositChannelExpiryTime: number;
   isDepositChannelExpired: boolean;
   depositChannelOpenedThroughBackend: boolean;
+  affiliateBrokers?: { account: string; commissionBps: number }[];
 }
 
 type BroadcastRequested = {