+++ packages/swap/src/event-handlers/__tests__/swapScheduled.test.ts
@@ -100,7 +100,44 @@ describe(swapScheduled, () => {
         await swapScheduled({
           prisma: client,
           block: swapScheduledDotDepositChannelBrokerCommissionMock.block as any,
-          event: swapScheduledDotDepositChannelBrokerCommissionMock.eventContext.event as any,
+          event: {
+            ...swapScheduledDotDepositChannelBrokerCommissionMock.eventContext.event,
+            args: {
+              ...swapScheduledDotDepositChannelBrokerCommissionMock.eventContext.event.args,
+              brokerCommission: 5000000000,
+            },
+          },
+        });
+      });
+
+      const swap = await prisma.swap.findFirstOrThrow({
+        where: { swapDepositChannelId: dotSwapDepositChannel.id },
+        include: { fees: true },
+      });
+
+      expect(swap.depositAmount.toString()).toEqual(
+        swapScheduledDotDepositChannelMock.eventContext.event.args.depositAmount,
+      );
+      expect(swap).toMatchSnapshot({
+        id: expect.any(BigInt),
+        createdAt: expect.any(Date),
+        updatedAt: expect.any(Date),
+        swapDepositChannelId: expect.any(BigInt),
+        fees: [
+          {
+            id: expect.any(BigInt),
+            swapId: expect.any(BigInt),
+          },
+        ],
+      });
+    });
+
+    it('stores a new swap with a broker fee', async () => {
+      await prisma.$transaction(async (client) => {
+        await swapScheduled({
+          prisma: client,
+          block: swapScheduledDotDepositChannelBrokerCommissionMock.block as any,
+          event: swapScheduledDotDepositChannelBrokerCommissionMock.eventContext.event,
         });
       });