+++ packages/swap/src/event-handlers/swapScheduled.ts
@@ -17,16 +17,24 @@ const vaultSwapOrigin = z.object({
   txHash: z.string(),
 });
 
-const swapScheduledArgs = z.object({
-  swapId: u64,
-  sourceAsset: internalAssetEnum,
-  depositAmount: u128,
-  destinationAsset: internalAssetEnum,
-  destinationAddress: encodedAddress,
-  origin: z.union([depositChannelSwapOrigin, vaultSwapOrigin]),
-  swapType: swapTypeSchema,
-  brokerCommission: u128.optional(),
-});
+const swapScheduledArgs = z
+  .object({
+    swapId: u64,
+    sourceAsset: internalAssetEnum,
+    depositAmount: u128,
+    destinationAsset: internalAssetEnum,
+    destinationAddress: encodedAddress,
+    origin: z.union([depositChannelSwapOrigin, vaultSwapOrigin]),
+    swapType: swapTypeSchema,
+    // < v1.4.0
+    brokerCommission: u128.nullish(),
+    // >= v1.4.0
+    brokerFee: u128.nullish(),
+  })
+  .transform(({ brokerCommission, brokerFee, ...rest }) => ({
+    ...rest,
+    brokerFee: brokerFee ?? brokerCommission,
+  }));
 
 export type SwapScheduledEvent = z.input<typeof swapScheduledArgs>;
 
@@ -43,7 +51,7 @@ export default async function swapScheduled({
     destinationAddress,
     origin,
     swapType,
-    brokerCommission: brokerCommissionAmount,
+    brokerFee: brokerFeeAmount,
   } = swapScheduledArgs.parse(event.args);
 
   const newSwapData = {
@@ -52,12 +60,12 @@ export default async function swapScheduled({
     swapInputAmount: depositAmount.toString(),
     nativeId: swapId,
     depositReceivedAt: new Date(block.timestamp),
-    fees: brokerCommissionAmount
+    fees: brokerFeeAmount
       ? {
           create: {
             type: 'BROKER' as const,
             asset: sourceAsset,
-            amount: brokerCommissionAmount.toString(),
+            amount: brokerFeeAmount.toString(),
           },
         }
       : undefined,