+++ CHANGELOG.md
@@ -17,6 +17,11 @@ which it is deprecated.
 
 ## Unreleased
 
+### Added
+
+- `SwapSDK.prototype.getStatus` will return the list of affiliate brokers for a
+  deposit channel if it was opened with affiliates.
+
 ### Changed
 
 - `SwapSDK.prototype.executeSwap` allows to not wait for transaction inclusion